ai/zhipuai/requirements.txt
PyJWT
requests
python-dotenv